refactor: import type utils from `utils/type`
