+++ apps/calendar/src/controller/panel.ts
@@ -1,7 +1,6 @@
-import isUndefined from 'tui-code-snippet/type/isUndefined';
-
 import { Direction } from '@src/controller/layout';
 import { getElementRect as getRect } from '@src/utils/dom';
+import { isUndefined } from '@src/utils/type';
 
 export type Styles = Record<string, string | number>;
 export type PanelElementRectMap = Record<string, PanelRect>;
