+++ apps/calendar/src/template/default.tsx
@@ -1,13 +1,12 @@
 import { h } from 'preact';
 
-import isUndefined from 'tui-code-snippet/type/isUndefined';
-
 import { cls } from '@src/helpers/css';
 import { capitalizeDayName, getDayName } from '@src/helpers/dayName';
 import EventModel from '@src/model/eventModel';
 import TZDate from '@src/time/date';
 import { isSameDate, leadingZero, toFormat } from '@src/time/datetime';
 import { stripTags } from '@src/utils/dom';
+import { isUndefined } from '@src/utils/type';
 
 import { EventCategory } from '@t/events';
 import {
