+++ apps/calendar/src/template/index.ts
@@ -1,10 +1,9 @@
-import isNumber from 'tui-code-snippet/type/isNumber';
-
 import { cls } from '@src/helpers/css';
 import EventUIModel from '@src/model/eventUIModel';
 import { templates } from '@src/template/default';
 import TZDate from '@src/time/date';
 import { toFormat } from '@src/time/datetime';
+import { isNumber } from '@src/utils/type';
 
 import { GridUIModel } from '@t/grid';
 import { Template, TemplateConfig } from '@t/template';
