+++ apps/calendar/src/utils/dom.ts
@@ -1,8 +1,8 @@
 import getMousePosition from 'tui-code-snippet/domEvent/getMousePosition';
 import getTarget from 'tui-code-snippet/domEvent/getTarget';
-import isString from 'tui-code-snippet/type/isString';
 
 import { noop } from '@src/utils/noop';
+import { isString } from '@src/utils/type';
 
 const CSS_AUTO_REGEX = /^auto$|^$|%/;
